sport DAO delete sport unit test

+++ test/com/asiangames2018/TestCountryDAO.java
@@ -10,12 +10,12 @@ import com.asiangames2018.dao.AsianGamesDAO;
 import com.asiangames2018.entity.Country;
 
 public class TestCountryDAO {
-   
+
     /**
-     * To test insert, delete, and update country, you
-     * have to test them in order of insert, update, and delete.
+     * To test insert, delete, and update country, you have to test them in
+     * order of insert, update, and delete.
      */
-    
+
     @Test
     public void testListAllCountries() {
 	AsianGamesDAO dao = new AsianGamesDAO();
@@ -26,14 +26,14 @@ public class TestCountryDAO {
 	    System.out.println(c.getCountryId());
 	}
     }
-    
+
     @Test
     public void testIsCountryExist() {
 	AsianGamesDAO dao = new AsianGamesDAO();
 	assertTrue(dao.isCountryExist(new Country("PHI", "Philippines")));
 	assertFalse(dao.isCountryExist(new Country("XYZ", "Nowhere")));
     }
-    
+
     @Test
     public void testInsertCountry() {
 	AsianGamesDAO dao = new AsianGamesDAO();
@@ -41,7 +41,7 @@ public class TestCountryDAO {
 	dao.insertCountry(can);
 	assertTrue(dao.isCountryExist(can));
     }
-    
+
     @Test
     public void testUpdateCountry() {
 	AsianGamesDAO dao = new AsianGamesDAO();
@@ -50,13 +50,17 @@ public class TestCountryDAO {
 	dao.updateCountry(can);
 	assertTrue(dao.isCountryExist(can));
     }
-    
+
     @Test
     public void testDeleteCountry() {
 	AsianGamesDAO dao = new AsianGamesDAO();
 	Country can = new Country("CAN", "Canada");
-	dao.deleteCountry(can);
-	assertFalse(dao.isCountryExist(can)); // this country must not exist after deletion
+	if (dao.isCountryExist(can)) {
+	    dao.deleteCountry(can);
+	    assertFalse(dao.isCountryExist(can));
+	}
+	dao.insertCountry(can);
+	assertTrue(dao.isCountryExist(can));
     }
-    
+
 }
